requirements.txt
Flask>=2.0
PyJWT>=2.0
Flask-Bcrypt>=1.0.1
Flask-Cors>=3.0.10
python-dotenv>=0.20.0
pymongo>=4.0
